Nico

+++ memoria/include/configuracion_memoria.h
@@ -6,7 +6,7 @@
     char* puerto_escucha;            //Puerto al cual se escuchara la conexion de modulo
     uint16_t tam_memoria;               //Tamaño expresado en bytes del espacio de usuario de la memoria.
     uint16_t tam_pagina;                //Tamaño de las paginas en bytes
-    uint16_t retardo_respuesta;         //Tiempo en milisegundos que se deberá esperar antes de responder a las solicitudes de CPU.
+    uint8_t retardo_respuesta;         //Tiempo en milisegundos que se deberá esperar antes de responder a las solicitudes de CPU.
     char* path_instrucciones;           //Carpeta donde se encuentran los archivos de pseudocódigo.
 
 }APP_config;
